Base Files for Game Engine
Updated the file system of the project (again sorry). Added a separate folder for the Game Engine and all of it's files, as I'm reading that is the correct thing to do so as not to get all confused in the main application, source files, and classes of the project. Added are some of the base files that we will need, including Render and Physics, being the main 2 that will be worked on in the Game Engine by Chris and John respectively. I will continue to research proper game enignes, best practices, and the necessities of a base game engine, to get something working for the end of the week. It will involve a lot of splitting code so bear with me. Again please try and branch into the features/ that you need to. As it stands, John and Chris will need to branch from feature/gameengine to begin. Have a good weekend all!