Engines Start
Setup the engine and the base files for the engine. Renamed a lot of the folders to remove the spacing in them. Following the procedure for all the files, but nothing has been implemented as of yet. Working to start getting this moved into the main project and test if there is initialization and deconstruction properly.